+++ src/test/java/com/programmers/smrtstore/domain/product/application/ProductServiceTest.java
@@ -3,6 +3,11 @@ package com.programmers.smrtstore.domain.product.application;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import com.programmers.smrtstore.domain.product.application.dto.req.CreateProductOptionRequest;
+import com.programmers.smrtstore.domain.product.application.dto.req.CreateProductRequest;
+import com.programmers.smrtstore.domain.product.application.dto.req.UpdateProductOptionRequest;
+import com.programmers.smrtstore.domain.product.application.dto.req.UpdateProductRequest;
+import com.programmers.smrtstore.domain.product.application.dto.res.ProductResponse;
 import com.programmers.smrtstore.domain.product.domain.entity.Category;
 import com.programmers.smrtstore.domain.product.domain.entity.OptionTag;
 import com.programmers.smrtstore.domain.product.domain.entity.Product;
@@ -10,11 +15,6 @@ import com.programmers.smrtstore.domain.product.domain.entity.ProductOption;
 import com.programmers.smrtstore.domain.product.exception.ProductException;
 import com.programmers.smrtstore.domain.product.infrastructure.ProductJPARepository;
 import com.programmers.smrtstore.domain.product.infrastructure.ProductOptionJPARepository;
-import com.programmers.smrtstore.domain.product.presentation.dto.req.CreateProductOptionRequest;
-import com.programmers.smrtstore.domain.product.presentation.dto.req.CreateProductRequest;
-import com.programmers.smrtstore.domain.product.presentation.dto.req.UpdateProductOptionRequest;
-import com.programmers.smrtstore.domain.product.presentation.dto.req.UpdateProductRequest;
-import com.programmers.smrtstore.domain.product.presentation.dto.res.ProductResponse;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.List;
